backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pypdf==3.17.0
Pillow==10.4.0
python-dotenv==1.0.0
pdf2image==1.16.3
reportlab==4.0.7
weasyprint==60.1
pytesseract==0.3.10
PyMuPDF==1.23.8
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
python-pptx==0.6.23
requests==2.31.0
pdf2docx==0.5.8
pdfservices-sdk==4.0.0
